# (rc) drastically improve performance when verifying multiple configurations