--- a/releasenotes/notes/rc-performance-multi-config-e1530e77fdc05dd2.yaml
+++ b/releasenotes/notes/rc-performance-multi-config-e1530e77fdc05dd2.yaml
@@ -1,0 +1,11 @@
+# Each section from every release note are combined when the
+# CHANGELOG.rst is rendered. So the text needs to be worded so that
+# it does not depend on any information only available in another
+# section. This may mean repeating some details, but each section
+# must be readable independently of the other.
+#
+# Each section note must be formatted as reStructuredText.
+---
+enhancements:
+  - |
+    The performance of Remote Config has been significantly improved when large amounts of configurations are received.