… comp/core/log/go.sum
… go.sum
… pkg/config/remote/go.sum
… pkg/remoteconfig/state/go.sum
… pkg/trace/go.sum