### What does this PR do ?

This PR uses the batch operations introduced in https://github.com/DataDog/go-tuf/pull/67 to improve the verification performance of the RC service. It is aimed at scaling better when many configurations are received by the agent. 

To give an example, in a scenario where the agent is receiving 128 configs, the max we have seen in real life scenarios, we went from 16000ms to 13ms on my CPU, a `99.18%` decrease.

It should fix the two issues we have seen in prod:
- Remote Config verification taking up to 10% of the CPU used by the agent in some clusters
- Agents with 100s of Dynamic Instrumentation configs being almost always working on verification and not answering the connected tracers

Raw results:
```
benchmark                                 old ns/op      new ns/op     delta
BenchmarkVerify/verify-1-configs-12       2366747        2341795       -1.05%
BenchmarkVerify/verify-2-configs-12       5135915        2469946       -51.91%
BenchmarkVerify/verify-4-configs-12       11120302       2637195       -76.28%
BenchmarkVerify/verify-8-configs-12       25292809       2975712       -88.23%
BenchmarkVerify/verify-16-configs-12      60258463       3743521       -93.79%
BenchmarkVerify/verify-32-configs-12      170096629      5154561       -96.97%
BenchmarkVerify/verify-64-configs-12      515783400      8012378       -98.45%
BenchmarkVerify/verify-128-configs-12     1699531400     13946932      -99.18%

benchmark                                 old allocs     new allocs     delta
BenchmarkVerify/verify-1-configs-12       7014           7014           +0.00%
BenchmarkVerify/verify-2-configs-12       14831          7554           -49.07%
BenchmarkVerify/verify-4-configs-12       32990          8603           -73.92%
BenchmarkVerify/verify-8-configs-12       79419          10683          -86.55%
BenchmarkVerify/verify-16-configs-12      213570         14889          -93.03%
BenchmarkVerify/verify-32-configs-12      643968         23285          -96.38%
BenchmarkVerify/verify-64-configs-12      2153727        39994          -98.14%
BenchmarkVerify/verify-128-configs-12     7758737        73306          -99.06%

benchmark                                 old bytes     new bytes     delta
BenchmarkVerify/verify-1-configs-12       376539        376411        -0.03%
BenchmarkVerify/verify-2-configs-12       799136        411481        -48.51%
BenchmarkVerify/verify-4-configs-12       1815733       483825        -73.35%
BenchmarkVerify/verify-8-configs-12       4397357       600218        -86.35%
BenchmarkVerify/verify-16-configs-12      13190050      929495        -92.95%
BenchmarkVerify/verify-32-configs-12      43301026      1563221       -96.39%
BenchmarkVerify/verify-64-configs-12      151027728     2783854       -98.16%
BenchmarkVerify/verify-128-configs-12     572859936     5325649       -99.07%
```

Note: I expect even better results in real life scenarios because we have much more TUF metadata than in the benchmark test where there's only one top targets and no delegated targets.

### Followup work

This PR should address all known issues with RC performance in the agent. There's still room for improvements by improving JSON operations in the TUF library but given that it's not trivial (it uses a custom canonical JSON lib) and we haven't seen issues with low amounts of configs let's wait for now.